moved to OZ finally + added virtual assets counting

+++ src/contracts/interfaces/IStakeToken.sol
@@ -6,7 +6,7 @@ import {IERC4626} from 'openzeppelin-contracts/contracts/interfaces/IERC4626.sol
 interface IStakeToken is IERC4626 {
   struct CooldownSnapshot {
     /// @notice Amount of tokens available for withdrawal
-    uint224 amount;
+    uint192 amount;
     /// @notice Time to unlock funds for withdrawal
     uint32 timestamp;
   }
