+++ tests/ERC4626.t.sol
@@ -32,7 +32,7 @@ contract ERC4626Tests is StakeTestBase {
   }
 
   // Due to default 1e18 exchange rate there's no rounding here at all, so I checked these values striclty
-  function test_previewFunctions(uint224 assets) public view {
+  function test_previewFunctions(uint192 assets) public view {
     uint256 shares = stakeToken.convertToShares(assets);
 
     uint256 previewDeposit = stakeToken.previewDeposit(assets);
@@ -48,7 +48,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(previewRedeem, assets);
   }
 
-  function test_deposit(uint224 amountToStake) public {
+  function test_deposit(uint192 amountToStake) public {
     vm.assume(amountToStake > 0);
 
     uint256 shares = _deposit(amountToStake, user, user);
@@ -60,7 +60,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), shares);
   }
 
-  function test_depositToSomeone(uint224 amountToStake) public {
+  function test_depositToSomeone(uint192 amountToStake) public {
     vm.assume(amountToStake > 0);
 
     uint256 shares = _deposit(amountToStake, user, someone);
@@ -72,7 +72,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(someone), shares);
   }
 
-  function test_mint(uint224 amountOfShares) public {
+  function test_mint(uint192 amountOfShares) public {
     vm.assume(amountOfShares > 0);
 
     uint256 amountToStake = stakeToken.convertToAssets(amountOfShares);
@@ -87,7 +87,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), amountOfShares);
   }
 
-  function test_mintToSomeone(uint224 amountOfShares) public {
+  function test_mintToSomeone(uint192 amountOfShares) public {
     vm.assume(amountOfShares > 0);
 
     uint256 amountToStake = stakeToken.convertToAssets(amountOfShares);
@@ -102,7 +102,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(someone), amountOfShares);
   }
 
-  function test_maxWithdraw(uint224 amountToStake) public {
+  function test_maxWithdraw(uint192 amountToStake) public {
     vm.assume(amountToStake > 0);
 
     deal(address(underlying), user, amountToStake);
@@ -124,7 +124,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(assetsAvailable, amountToStake);
   }
 
-  function test_maxRedeem(uint224 amountToStake) public {
+  function test_maxRedeem(uint192 amountToStake) public {
     vm.assume(amountToStake > 0);
 
     uint256 shares = _deposit(amountToStake, user, user);
@@ -144,7 +144,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(sharesAvailable, shares);
   }
 
-  function test_redeem(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_redeem(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -166,7 +166,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_redeemToSomeone(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_redeemToSomeone(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -188,7 +188,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_redeemWithApprove(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_redeemWithApprove(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -213,7 +213,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_redeemWithoutApprove(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_redeemWithoutApprove(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -239,7 +239,7 @@ contract ERC4626Tests is StakeTestBase {
     stakeToken.redeem(sharesToRedeem, someone, user);
   }
 
-  function test_redeemMoreThanHave(uint224 amountStaked) public {
+  function test_redeemMoreThanHave(uint192 amountStaked) public {
     vm.assume(amountStaked > 0);
 
     uint256 shares = _deposit(amountStaked, user, user);
@@ -261,7 +261,7 @@ contract ERC4626Tests is StakeTestBase {
     stakeToken.redeem(shares + 1, user, user);
   }
 
-  function test_withdraw(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_withdraw(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -285,7 +285,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_withdrawToSomeone(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_withdrawToSomeone(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -309,7 +309,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_withdrawWithApprove(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_withdrawWithApprove(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -336,7 +336,7 @@ contract ERC4626Tests is StakeTestBase {
     assertEq(stakeToken.balanceOf(user), stakeToken.convertToShares(amountStaked - amountRedeemed));
   }
 
-  function test_withdrawWithoutApprove(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_withdrawWithoutApprove(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
 
@@ -363,7 +363,7 @@ contract ERC4626Tests is StakeTestBase {
     stakeToken.withdraw(amountRedeemed, someone, user);
   }
 
-  function test_withdrawMoreThanHave(uint224 amountStaked) public {
+  function test_withdrawMoreThanHave(uint192 amountStaked) public {
     vm.assume(amountStaked > 0);
 
     _deposit(amountStaked, user, user);
@@ -385,7 +385,7 @@ contract ERC4626Tests is StakeTestBase {
     stakeToken.withdraw(uint256(amountStaked) + 1, user, user);
   }
 
-  function test_events(uint224 amountStaked, uint224 amountRedeemed) public {
+  function test_events(uint192 amountStaked, uint192 amountRedeemed) public {
     vm.assume(amountStaked > 0);
     vm.assume(amountRedeemed != 0 && amountRedeemed <= amountStaked);
